Restructuring utility for codenames and versioning
