+++ himalaya/internal/collection/entry/metadata.go
